moved logger to src folder and removed tests
