--- Examples/LoggerSkeleton/main.cpp
@@ -1,13 +1,0 @@
-#include <QCoreApplication>
-#include "LoggerSkeleton.h"
-
-
-
-int main(int argc, char **argv)
-{
-	QCoreApplication app(argc, argv);
-	Logger::Init("log1file.log");
-	qDebug() << "hello" << "\n";
-	qInfo() << "hello_info" << "\n";
-	return app.exec();
-}
